Revert changes to fork choice reference tests.
