+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/teku/reference/phase0/fork_choice/ForkChoiceTestExecutor.java
@@ -39,16 +39,19 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import tech.pegasys.teku.core.ForkChoiceUtil;
+import tech.pegasys.teku.core.StateTransition;
+import tech.pegasys.teku.core.results.BlockImportResult;
+import tech.pegasys.teku.datastructures.attestation.ValidateableAttestation;
 import tech.pegasys.teku.datastructures.blocks.SignedBeaconBlock;
 import tech.pegasys.teku.datastructures.operations.Attestation;
 import tech.pegasys.teku.datastructures.state.BeaconState;
 import tech.pegasys.teku.datastructures.state.BeaconStateImpl;
+import tech.pegasys.teku.datastructures.util.AttestationProcessingResult;
 import tech.pegasys.teku.datastructures.util.SimpleOffsetSerializer;
 import tech.pegasys.teku.ethtests.finder.TestDefinition;
+import tech.pegasys.teku.protoarray.ForkChoiceStrategy;
+import tech.pegasys.teku.protoarray.ProtoArrayForkChoiceStrategy;
 import tech.pegasys.teku.reference.phase0.TestExecutor;
-import tech.pegasys.teku.reference.phase0.fork_choice.implementatations.ForkChoiceProcessor;
-import tech.pegasys.teku.reference.phase0.fork_choice.implementatations.OrigForkChoiceProcessor;
-import tech.pegasys.teku.reference.phase0.fork_choice.implementatations.StoreBackedForkChoiceProcessor;
 import tech.pegasys.teku.storage.client.MemoryOnlyRecentChainData;
 import tech.pegasys.teku.storage.client.RecentChainData;
 import tech.pegasys.teku.storage.store.UpdatableStore;
@@ -168,19 +171,20 @@ public class ForkChoiceTestExecutor implements TestExecutor {
     runForkChoiceTests(genesis, steps, testName, protoArrayFC);
   }
 
-  @ParameterizedTest(name = "{index}. (protoarray={3}) {2}")
+  @ParameterizedTest(name = "{index}.{2} fork choice test")
   @MethodSource("loadForkChoiceTests")
   void runForkChoiceTests(
       BeaconState genesis, List<Object> steps, String testName, boolean protoArrayFC) {
+    StateTransition st = new StateTransition();
 
     EventBus eventBus = new EventBus();
     RecentChainData storageClient = MemoryOnlyRecentChainData.create(eventBus);
     storageClient.initializeFromGenesis(genesis);
 
-    ForkChoiceProcessor forkChoiceProcessor =
+    ForkChoiceStrategy forkChoiceStrategy =
         protoArrayFC
-            ? new StoreBackedForkChoiceProcessor(storageClient)
-            : new OrigForkChoiceProcessor(storageClient);
+            ? ProtoArrayForkChoiceStrategy.create(storageClient.getStore())
+            : new OrigForkChoiceStrategy(storageClient.getStore());
 
     @SuppressWarnings("ModifiedButNotUsed")
     List<SignedBeaconBlock> blockBuffer = new ArrayList<>();
@@ -188,8 +192,9 @@ public class ForkChoiceTestExecutor implements TestExecutor {
     List<Attestation> attestationBuffer = new ArrayList<>();
 
     for (Object step : steps) {
-      blockBuffer.removeIf(forkChoiceProcessor::processBlock);
-      attestationBuffer.removeIf(forkChoiceProcessor::processAttestation);
+      blockBuffer.removeIf(block -> processBlock(st, storageClient, block, forkChoiceStrategy));
+      attestationBuffer.removeIf(
+          attestation -> processAttestation(st, storageClient, attestation, forkChoiceStrategy));
       if (step instanceof UnsignedLong) {
         UpdatableStore.StoreTransaction transaction = storageClient.startStoreTransaction();
         while (ForkChoiceUtil.get_current_slot(transaction).compareTo((UnsignedLong) step) < 0) {
@@ -202,11 +207,11 @@ public class ForkChoiceTestExecutor implements TestExecutor {
             ((SignedBeaconBlock) step).getMessage().getBody().getAttestations()) {
           attestationBuffer.add(attestation);
         }
-        if (!forkChoiceProcessor.processBlock((SignedBeaconBlock) step)) {
+        if (!processBlock(st, storageClient, (SignedBeaconBlock) step, forkChoiceStrategy)) {
           blockBuffer.add((SignedBeaconBlock) step);
         }
       } else if (step instanceof Attestation) {
-        if (!forkChoiceProcessor.processAttestation((Attestation) step)) {
+        if (!processAttestation(st, storageClient, (Attestation) step, forkChoiceStrategy)) {
           attestationBuffer.add((Attestation) step);
         }
       } else if (step instanceof Map) {
@@ -234,7 +239,9 @@ public class ForkChoiceTestExecutor implements TestExecutor {
             case "head":
               {
                 Bytes32 root = Bytes32.fromHexString((String) e.getValue());
-                Bytes32 head = forkChoiceProcessor.processHead();
+                UpdatableStore.StoreTransaction transaction = storageClient.startStoreTransaction();
+                Bytes32 head = forkChoiceStrategy.findHead(transaction);
+                transaction.commit(() -> {}, "Failed to persist validator vote changes.");
                 assertEquals(
                     root,
                     head,
@@ -264,6 +271,41 @@ public class ForkChoiceTestExecutor implements TestExecutor {
       } else {
         throw new IllegalArgumentException();
       }
+    }
+  }
+
+  private boolean processAttestation(
+      StateTransition st,
+      RecentChainData storageClient,
+      Attestation step,
+      ForkChoiceStrategy strategy) {
+    UpdatableStore.StoreTransaction transaction = storageClient.startStoreTransaction();
+    AttestationProcessingResult attestationProcessingResult =
+        ForkChoiceUtil.on_attestation(
+            transaction, ValidateableAttestation.fromAttestation(step), st, strategy);
+    if (attestationProcessingResult.isSuccessful()) {
+      transaction.commit().join();
+      return true;
+    } else {
+      return false;
+    }
+  }
+
+  private boolean processBlock(
+      StateTransition st,
+      RecentChainData storageClient,
+      SignedBeaconBlock block,
+      ForkChoiceStrategy forkChoiceStrategy) {
+    UpdatableStore.StoreTransaction transaction = storageClient.startStoreTransaction();
+    BlockImportResult blockImportResult =
+        ForkChoiceUtil.on_block(transaction, block, st, forkChoiceStrategy);
+    if (blockImportResult.isSuccessful()) {
+      transaction.commit().join();
+      forkChoiceStrategy.onBlock(
+          block.getMessage(), blockImportResult.getBlockProcessingRecord().get().getPostState());
+      return true;
+    } else {
+      return false;
     }
   }
